+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * require user to confirm password before disabling encryption
 * when enabling encryption from settings page, display a second password confirmation field to avoid mistyping password.
 * allow user to change the current unit price when editing a tradeable asset.
+* add a close button to the "Connect your storage" popup when not connected to a backend, so user can quickly disable cloud sync.
 
 ## 1.0.0-beta (2019-12-10)